Experiment to see if custom fonts can be loaded sooner
When a new device accesses the UI for the first time, the UI can be
rendered with Times New Roman while it waits for the album/track/etc
lists to be retrieved. It would be nice if the custom fonts could be
loaded sooner. This is just a quick experiment to see if it helps.