Fix regression breaking favorite playback; fix scroll on modal display
Re-enable replacing playlist with favorites. Also add tooltip showing
that this capability is disabled when the streamer is in standby.

Remove `<ScrollRestoration>` from `<RootLayout>` as it was forcing
scroll-to-top behavior when a modal (such as lyrics or hotkeys) was
shown, and it didn't seem to be doing anything useful otherwise.

Notes:

* Need to think about a standardized app-wide approach to showing the
  user when capabilities are disabled due to the streamer being in
  standby.
* Need to find a good way to restore scroll position when returning to
  large screens like Albums and Tracks.